Transformed into a package, including skeletton of docs. Reformatting.
