+++ .gitignore
@@ -14,7 +14,7 @@ deps/usr/
 deps/src/
 
 # Build artifacts for creating documentation generated by the Documenter package
-docs/build/
+docs/build*
 docs/site/
 
 # File generated by Pkg, the package manager, based on a corresponding Project.toml
