+++ Project.toml
@@ -1,0 +1,21 @@
+name = "CausticsEngineering"
+uuid = "e17b973c-3eae-4e30-9cf2-ee583e3dfd43"
+authors = ["Matt Ferraro", "Emmanuel Rialland"]
+version = "0.1.0"
+
+[deps]
+DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
+Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
+JuliaFormatter = "98e50ef6-434e-11e9-1051-2b60c6c9e899"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[compat]
+julia = "1"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
