+++ README.md
@@ -6,6 +6,11 @@ See the write-up [here](https://mattferraro.dev/posts/caustics-engineering)!
 
 # To Run
 
+To run the cat example from the blogpost, run line by line from ` src/scratchpad.jl`.
+
+_OR_ from the command line.
+
 ```
-julia create_mesh.jl cat_posing.jpg
+julia ./run.jl"
 ```
+The image file is currently hard-coded.
