+++ src/CausticsEngineering.jl
@@ -1,0 +1,13 @@
+module CausticsEngineering
+
+using DocStringExtensions
+
+using Images, Plots
+gr()
+
+include("utilities.jl")
+include("create_mesh.jl")
+
+export main, engineer_caustics
+
+end
