+++ test/runtests.jl
@@ -1,0 +1,6 @@
+using CausticsEngineering
+using Test
+
+@testset "CausticsEngineering.jl" begin
+    # Write your tests here.
+end
